Don't compile/link rasterizer.cpp/matrixMath.cpp unless "scons raster=1" used
This partially reverts r6585 "Always link with gfx_newRenderPlan code, removing scons raster option"

Reduces game.exe size by ~43KB on my machine. Maybe it'll cut much more on the
Windows nightly build machine, which is linking in 1MB of unnecessary C++
libraries for unknown reasons. And if so, the difference could even be a fix for
gh#38 Virus scanners raise false alarm on game.exe when a game icon is embedded

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11480 7d344553-34f0-0310-a9b1-970ce8f1c3a2