You can now copy-paste between browsers for different spriteset types.
This partially undoes the previous revision because now the number of frames and
the frame IDs are NOT copied, they are reset to defaults to ensure you can't
mess them up, and then the copy_buffer is pasted on top. Frame size is still
copied.
That change will be reverted once we support variable frame numbers.

Help page already claimed you could copy-paste between editors.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11492 7d344553-34f0-0310-a9b1-970ce8f1c3a2