plankmenu: left/right at the end of a row now wraps around
Actually what it does is go to the previous/next plank in tree traversal order,
so it depends on that order being the order you expect. (In fact it was not
in the spriteset browser, so I needed to fix that).

When we expose plankmenus to users we'll probably want to have an option to
disable this behaviour.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11502 7d344553-34f0-0310-a9b1-970ce8f1c3a2