Fix SpriteSetBrowser bug which caused inconsistent defpals in defpal#.bin & .rgfx
I added a check for defpal inconsistency when entering the browser, if any is
found it's fixed (preferring defpal#.bin) and a warning shown.

Also add Ctrl-R debug key to force a reload of the script cache,

The bug was added in r11542/r11543. I forgot to write a line of code!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11555 7d344553-34f0-0310-a9b1-970ce8f1c3a2