Improve display of spread-targetted attacks in the OOB target picker
Rather than just highlighting all four hero slots, it now actually shows
which heroes will be hit by the spread, which might not be all of them.

For spread attacks it showed a flashing background which was hideous and painful
(depending on UI colours).  Highlight heroes with flashing text instead.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11604 7d344553-34f0-0310-a9b1-970ce8f1c3a2