Rewrite Palette16 cache. Just load whole .pal instead.
The palette cache was VERY slow especially in large games due to the heavy use
of strings, and even a trimpath() call. Most of the time was spent locking and
unlocking the FB string mutex. I didn't realise that operations on FB strings
in a multi-threaded program are so slow.

Creating a sprite slice is now faster depending on how many palettes were in
the cache, e.g. over 10x faster in jade.rpg.

As well as being simpler than the old cache, now all Palette16's have a refcount
(it is never NOREFC) so I could remove some kludges in render_text.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11605 7d344553-34f0-0310-a9b1-970ce8f1c3a2