gamestate_party_from_reload now calls addhero instead of duplicating initialisation
This fixes a bug in r11689: hero slices weren't loaded whn loading a game.
It's also far better and cleaner.

Importantly, I also removed the 'default' directives with that function, since
it's no longer necessary for any of the data to actually be read from
.rsav. This means that in future we can avoid writing hero data to .rsav if it's
the same as the hero definition data, allowing .rpg hero edits to affect saved
games!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11692 7d344553-34f0-0310-a9b1-970ce8f1c3a2