Use curmasterpal() instead of master() in a few places
This fixes glitches if a script changes the master palette but hasn't called
fadescreenin/updatepalette when drawing e.g. transrects or minimaps.
That's pretty unlikely, but especially when we allow 32-bit color mode, using
master() would cause transrects to be drawn wrong when the screen is faded.

Also changed findrgb to use nearcolor_fast instead of nearcolor (effectively
switches from master() to curmasterpal()).

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11771 7d344553-34f0-0310-a9b1-970ce8f1c3a2