Whoops, due to an FB bug you can't pass around curmasterpal() outside allmodex.bas!
Revert part of the last commit to fix crashes.

Since curmasterpal is a static array, when "curmasterpal()" is passed to a sub
FB creates temp array descriptor to pass. But this only works inside
allmodex.bas, in other modules an invalid blank descriptor is used.
I assumed that bug is well known but I couldn't find it in FB's bugtrackers.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11772 7d344553-34f0-0310-a9b1-970ce8f1c3a2