Fix verify_map_size scrambling pass/foemap or adding garbage, deleting zonemap
If the pass/foemap width changed, each row would get shifted over, like the old
broken map resize feature! If the pass/foemap was extended, the extra space was
filled with garbage. And the zonemap was just deleted instead of resized.

Use the existing functions for resizing map data instead of REALLOCATE nonsense.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11791 7d344553-34f0-0310-a9b1-970ce8f1c3a2