Cleanup map resizing
Got rid of the messages that print during resize; resizing is instantaneous these days.

If you exit the map resize menu by pressing Enter but there's no change to the
size, don't resize. Preserves undo.

When importing a tilemap that's the wrong size, if overwriting the map don't
take the user to the resize menu, just resize.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11802 7d344553-34f0-0310-a9b1-970ce8f1c3a2