Save damage and level caps in RSAV only if they're modified.
Existing save-games will continue to have potentially out-of-date caps, and
loading and resaving won't fix that.

I bumped the RSAV version because if you save a game and then load it with an
older version, damage cap will change to none and level cap will change to 99.

Also, it turns out I forgot to bump the RSAV version last time I documented a
bump!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11827 7d344553-34f0-0310-a9b1-970ce8f1c3a2