Implement fading in & out in 32-bit color mode, but not fading between two palettes
For the common cases of fading out and back in to the same palette, fades in
32-bit color mode should behave identically to 8-bit ones! I had once thought
that impossible to achieve.

However there might be some slight differences when they are mixed with setpal,
I think one less frame may be sent to the backend or gif recorder at the start
of a fade.

Fading in to a different palette results in a sudden jump after the mis-fade.

It works by continuing to draw to the video page with a non-faded palette
(curmasterpal) even when the screen is apparently faded out: when faded_in=NO,
setvispage ignores vpage and draws the screen a solid color. That way, fading in
can redraw the previous vpage contents.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11906 7d344553-34f0-0310-a9b1-970ce8f1c3a2