Tricks to provide backcompat for fades ins in 32-bit mode
The main problem was that you couldn't fade between two master palettes. That's
now accomplished by calling the new setdrawpal function, redrawing the page
(with game logic partially paused), and fading to it with the new fadetopage
function. So "fade screen in" now does an implicit wait(1) under 32-bit mode if
fading to a new master palette, e.g. after "greyscale palette". I think it's so
unlikely to break any games that I'm not putting a big warning on the 24-bit
color depth option.

As well as "fade screen in" causing fadetopage to be called, non-scripted queued
screen fades now call setdrawpal, so scripts which modify the palette but don't
call "fade screen in" or "update palette" also work the same as before.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11917 7d344553-34f0-0310-a9b1-970ce8f1c3a2