Fix SerializeZoneTiles bug that caused spans of 256 tiles to be serialized wrong
(A span is a horizontal row of tiles that are all 0 (not in the zone) or 1 (in
the zone)).  So if you had a map of width exactly 256 and filled it with a zone
it would be written empty.

Also, if you resized a map adding >= 256 tiles to the left an overflow would
cause all the zonemaps to shift left by a multiple of 256.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12031 7d344553-34f0-0310-a9b1-970ce8f1c3a2