distribmenu: fix Windows packaging to include SDL2 instead of SDL1.2 dlls
However I did NOT update the Linux packaging for SDL2, because Linux builds
aren't compiling currently, so we still have the old gfx_sdl builds on the server.

I also changed it to download ohrrpgce-player-win-sdl2-wip.zip instead of
ohrrpgce-player-win-wip.zip from http://hamsterrepublic.com/ohrrpgce/nightly/ so
that ohrrpgce-player-win-wip.zip can continue to point to a
gfx_directx+sdl/music_sdl build so that older nightly builds continue to be able
to package for Windows.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12053 7d344553-34f0-0310-a9b1-970ce8f1c3a2