Fix graphics editors not checking whether the mouse is active (over the window)
If the mouse leaves the window all our backends continue to report a position
inside the window. So then the tile/sprite editors may try to movemouse as the
arrow keys are used, which broke if using gfx_sdl2 (see also next commit).

Also, no longer draw the mouse cursor in these editors if the mouse isn't
active. That's bothered me a long time!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12090 7d344553-34f0-0310-a9b1-970ce8f1c3a2