Official SDL2.dll version 2.0.12 build (released 2020-03-11)
The previous dll was 2.0.7, which was released 2017-10-23

In particular, this has many fixes for audio on Windows. 2.0.6 had switched to
the new WASAPI driver by default on Windows, which had a lot of bugs.

I'm a bit annoyed that the size increased (when zipped) from 398kB to 457kB,
since it was already far larger than SDL.dll

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12105 7d344553-34f0-0310-a9b1-970ce8f1c3a2