Fix FB 1.06-7 compiles, which I broke in the previous commit
Turns out that the breaking change to __fb_ctx was made in 2015 on FB's
emscripten branch but only merged into master a couple weeks ago. I didn't
notice this in the git log.

So this means any FB builds from pre-1.08 git previous to the emscripten merge
are now not supported. You will not get a warning. Recompile FB.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12148 7d344553-34f0-0310-a9b1-970ce8f1c3a2