scons: don't use "git svn info" at all on any OS, it's so slow
Instead, use the "git log --grep ..." fallback already used on Windows.

On my Linux machine, this reduces the time for running "scons unlump" with no
rebuild needed from 1.4s to 0.8s, which feels like a lot in practice.  Because
scons does things in parallel, the reduction for a no-rebuild "scons" is only
1.5s -> 1.3s.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12222 7d344553-34f0-0310-a9b1-970ce8f1c3a2