On Linux, print backtraces also on FB errors, and improve error messages
As well as printing the backtrace also decode the signal number to a name, and
print the module and function names in debugging builds, like it did before
r12249.

Also catch a few more signals.

Also change shift-tab-F6 to a null ptr access test.

And try to avoid a gfx_close() deadlock when quitting from a fatal error

Also, if DrMingw is in use (which is never, except when I test it) also log
backtraces on Windows when showerror_internal is called.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12265 7d344553-34f0-0310-a9b1-970ce8f1c3a2