Two changes to battle cursor movement
If you put four targets in a diamond, 45 degrees to each other, a 5th one in the
center could be untargetable! (Example in test.rpg). To fix this, switch to the
fallback cursor movement behaviour when angle >= 45 degrees instead of > 45.

Secondly, change the fallback behaviour from "pick the target with the minimum
angle" to "pick the nearest target (by L1 norm)". This reduces the tendency to
skip over targets that are in a roughly straight line when pressing arrows
mostly perpendicular to the line. However, skipping will still happen when the
line is close to a 45 degree angle.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12280 7d344553-34f0-0310-a9b1-970ce8f1c3a2