Use builtin font for help pages, showerror messages, multichoice (and yesno), and version text in Custom
The main reason is to be able to display "hróðvitnir", but making various
errors, prompts and help readable regardless of font is a good bonus.

Also fixed a wrapprint call in multichoice() with wrong args.

Ultimately I want to display all text in Custom, except for strings shown
in-game, using a different font than the game font, but that's a lot of work.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12332 7d344553-34f0-0310-a9b1-970ce8f1c3a2