Add a mechanism for checking version of support utilities and upgrading them as needed
This replaces the zip-version-3.0.txt kludge with a similar generic system,
which is to have an optional <appname>-version.txt next to the .exe containing
numeric and textual version codes. But instead of Custom creating or modifying
the -version.txt file, it will be included in the downloaded .zip, so it's
always kept correct when a support zip is extracted over the existing files.

Also, support/*-version.txt files are packaged with engine downloads that
include support utilities, to avoid unnecessary downloads.

Only utilities for which we care about the version code need a -version.txt file
in the .zip, and right now that's only zip.exe. I've updated zip.zip on HR.com
to include zip-version.txt. The old zip-version-3.0.txt if present is
ignored.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12366 7d344553-34f0-0310-a9b1-970ce8f1c3a2