Fix game window created too large on multimonitor X11. Fixed only with gfx_sdl2
The problem was that os_get_screen_size returned size of the whole desktop
rather than the main monitor so some percentage of that could be huge.

When using gfx_sdl2, use SDL2 functions to get the size.

I also wrote some code that queries X11 extension Xinerama to fix this bug
without gfx_sdl2, but it adds a dependency on libXinerama so I left it
commented. Even SDL2 can run without it.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12371 7d344553-34f0-0310-a9b1-970ce8f1c3a2