gfx_sdl2: workaround possible unexitable fullscreen problem on Windows in Custom
On Windows (both XP and 10), if the window client area is the same (or nearly
the same) size as the display and is centered and active, Windows will display
it above the taskbar -- the same as "desktop fullscreen".  This doesn't seem to
be SDL's doing.  Since additionally the window size doesn't get restored when
leaving fullscreen, unlike other OSes (and is maximised when fullscreened), you
become stuck fullscreened!  (To reproduce: run Custom, fullscreen, exit
fullscreen, then change window zoom in F9, which causes a window recentering if
this workaround is disabled.)

This is a very conservative fix for this issue, by simply not recentering
the window if it's as large as the display.

(Copied from r12423 on wip branch)

git-svn-id: https://rpg.hamsterrepublic.com/source/rel/hrodvitnir@12430 7d344553-34f0-0310-a9b1-970ce8f1c3a2