gfx_sdl2: restore previous window size when unfullscreening
On Windows, if the window is resizable then it gets maximised when fullscreened,
but doesn't restore the size automatically when exiting. But on xfce4 it does
restore.
I left this fix unconditioned on OS.

This is also a better fix for the possible stuck fullscreen problem on Windows
which I added a workaround in r12423 for, so I've removed that workaround to
keep the code cleaner.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12433 7d344553-34f0-0310-a9b1-970ce8f1c3a2