gfx_sdl2: fix window moving to 0,0 when unfullscreening if you change zoom while fullscreen on Windows
I originally had a much more complicated fix for this bug which I spent far too
long on which fixed the bug on Windows and caused an identical bug to occur on
Linux, which I gave up on fixing! I didn't realise that I hadn't actually
accomplished anything until it was time to write the commit message!

Also, actually call SDL_SetWindowSize immediately when restoring the size,
seemingly no reason not to

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12435 7d344553-34f0-0310-a9b1-970ce8f1c3a2