Global NPC editor was badly broken when entered directly from NPC placement mode
There are three ways to directly edit a global NPC other than via npcdef_editor
("Edit Global NPCs").

If entered this way, any edits were discarded.  And after exiting the editor,
local NPC graphics were loaded into the global NPC npc_img() array, resizing it
to the wrong size and causing various array OOB crashes. And of course one-time-
use tag checking was broken too.

All this was heralded by the comment that read
"(npcdata should be an element of st.map.npc_def() and when it isn't we would be using the other overload)"
which indicated that the function didn't work for global NPCs, but it was wrong
about the other overload

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12520 7d344553-34f0-0310-a9b1-970ce8f1c3a2