mapedit: pass global_npcdef_editor the real map insead of a dummy one
This allows the global NPC editor to show names of zones and some default
settings for the current map.  It also allows it to delete NPC instances when
the definition is deleted, but mapedit_sanity_check_npc_instances did this
anyway, so the only difference is that you're told how many instances you're
about to delete on the current map.  mapedit_sanity_check_npc_instances is still
needed.

Also, npcdef_editor and handle_npc_def_delete now use pool_id instead of
an is_global bool.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12522 7d344553-34f0-0310-a9b1-970ce8f1c3a2