Fix relump.exe/unlump.exe build error by compiling them without lto=1
LTO causes the current version of gcc on the nightly build machine machine to
fail an assert, although it worked with the previous version.

This resulted in ohrrpgce-win-installer-wip.exe and ohrrpgce-util.zip not
uploading, while other nightly builds were created and uploaded with
relump.exe. Yet another reason to centralise packaging in a single python
script.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12540 7d344553-34f0-0310-a9b1-970ce8f1c3a2