Implement template slices
A template slice is by default invisible, paused, and skipped over for hit
detection, ignored by its parent slice when laying out its children, etc.
However it is shown in the slice editor, and you can press F10 to toggle the
visibility of template slices therein.

Until it stabilises, the Template property can only be set in the slice editor
in privileged mode.

Template slices can be cloned with InstantiateTemplate() which removes the
Template property and puts the clone immediately before the template.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12555 7d344553-34f0-0310-a9b1-970ce8f1c3a2