reload: Errors while loading a doc are now shown with reporterr instead of quietly logged
The exception, for now, being "failed to open <file>" is quietly logged because
I vaguely remember seeing that still happen due to an oversight somewhere.

I think showing players errors on data corruption is better, as long as they are ignorable or once-off.

Added version of reporterr to common_base.bas.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12558 7d344553-34f0-0310-a9b1-970ce8f1c3a2