Improved error checking and handling when loading slice collections
SliceLoadFromNode/SliceLoadFromFile now return a bool to tell whether they were
successful, but they try to keep going if they hit an unrecognised slice type
and load the rest of the tree, then return false.

LoadSliceCollection now returns a Slice ptr instead of loading into an existing
slice. Has a new arg return_partially_loaded, which is currently always
YES. Better to try to recover from errors, especially when loading collections
in the slice editor!

This fixes a number of places that checked "if collection = null" to check for
error, which couldn't possibly succeed. And other cleanup to error checking.

I also removed a number of places that tried to set a slice to Fill and then
call RefreshSliceScreenPos before loading a collection into it, which did
nothing (the root slices of current builtin collections are all set to fill
anyway)

Also, converted some debugs into showbugs.

Also, when loading check for Class slices which aren't slSpecial.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12572 7d344553-34f0-0310-a9b1-970ce8f1c3a2