sliceedit: automatically clean forbidden slices when importing a collection
No longer opens a subeditor if there's a problem, which was very messy.
Also, finds and reports all forbidden slices instead of just the first.
I also implemented checking for special lookup code kindlimit conditions, but
left it commented out because it's easy to violate the kindlimitPLANKSELECTABLE,
probably better not to spam people with errors. Maybe add a "correctness check"
option in Settings menu instead.

I also changed the condition for when the forbidden slice search is done, I
don't think it was right.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12585 7d344553-34f0-0310-a9b1-970ce8f1c3a2