Introducing EditorKit, a framework for creating data editors
EditorKit is intended to be flexible enough to reimplement almost every existing
editor easily with minimal code. It's flexible because menu items are defined
with code, not data.

It's implemented on top of standardmenu and ModularMenu, which does create some
inflexibility, but they can be extended to overcome that.  It's based on how the
Debug Menu and General Game Data menu are implemented: you just Extend
EditorKit, add a define_items() method which lists the menu items, and write
some code to load/save data if needed and you're done.

Currently only basic data type editing is done (except floats), plus partial
set of directives config settings and RELOAD node datatypes. No special
support for switching records or lists of items yet.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12634 7d344553-34f0-0310-a9b1-970ce8f1c3a2