Modifying an ellipse slice caused the cliprect to be reset, for one tick.
This meant that if an ancestor of the ellipse was set to clip children then
the ellipse itself and any following descendents weren't clipped.

Having a global (actually TLS global) cliprect was a mistake; this isn't the
first time this sort of bug has happened. It ought to be per render-target
(Frame) instead.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12666 7d344553-34f0-0310-a9b1-970ce8f1c3a2