find_windows_helper_app returned "_" on Windows if the program wasn't found!
This lead to some weird errors, such as if hspeak.exe was missing. The problem
was worst when called with try_install=NO, since try_install did still work
which gave users a way around it. For example, this broke reading progressive
jpegs unless jpegtran was already installed.

I broke it in r12366.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12715 7d344553-34f0-0310-a9b1-970ce8f1c3a2