Fixes and a lot of simplification for get_valid_targ by splitting out is_foe_of
Combatants with Turncoat or Defector status (but not both) couldn't targets
themselves with attacks that could target Self.

Also, Dead Allies is no longer limited to the hero team and Dead Foes to the
enemy team. However, it still doesn't work as expected, because
attack_can_hit_dead also needs to be updated.

This simplification assumes that a target is either a foe or an ally, not both
and not neither.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12723 7d344553-34f0-0310-a9b1-970ce8f1c3a2