Comment out code for advancing time if there are no targettable enemies
The check in turn-based mode (turn_mode_time_passage) was completely broken in
two or three different ways, thankfully it did nothing at all. I moved its
commented out remnant to a better spot.

The check in active-time mode (battle_meters_can_advance) for hero menus &
targetting was equivalent to just checking whether it was the player's turn, so
I commented out that part. I think the difference between what it intended and
what it did is pretty small if it exists; I'm not sure.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12728 7d344553-34f0-0310-a9b1-970ce8f1c3a2