Finally implement menu "Max rows to display: Default" instead of using 18
I listed this as a bugfix, but it's all of feature, bugfix and change.
state.size is now calculated from the available space, taking into account the
menu position and alignment: added recalc_menu_size overload for MenuDef.

Quite a lot of code movement and cleanup was needed here.

This fixes a problem in popup_menu/multichoice where clamp_to_screen didn't work

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12745 7d344553-34f0-0310-a9b1-970ce8f1c3a2