Change draw_textured_background to a simpler, sparse, fuzzy pattern
The previous pattern caused really bad flashing when you moved a 50% fuzzy
rect around on top of it (such as a translucent menu). This fuzzy value was
choosen to avoid that and otherwise look decent.

But I'm not very happy with using fuzzy rects as backgrounds for previewing
in-game graphics.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12752 7d344553-34f0-0310-a9b1-970ce8f1c3a2