Try to avoid the pain of the render API's RGBPalette with masterpal_to_gfxpal kludge
I didn't want to remove RGBPalette because I think it'll be needed someday for
hardware accelerated rendering, but it's too obnoxious to use when the rest of
the codebase uses a RGBcolor FB array, and do all those expensive conversions.

Added unrollPalette16 so can use the render API with Palette16s

Deleted gfx_paletteCreate, can't imagine needing it.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12790 7d344553-34f0-0310-a9b1-970ce8f1c3a2