rasterizer: rewrite iteration across rows; fixes the glitches and far faster
Quads now (nearly always) draw without visible seams between the triangles,
because the tex/color at the center of each pixel (in the x coord, not yet for y
coords) is computed. The old code also computed the row length wrong.

Compute the increment for the TexCoord and/or Color for each pixel advanced,
rather than using lots of multiplications and divisions.  Added Color16 and
VertexPC16, VertexPTC16 types for this purpose.

Eliminated a few kludges. The minuteScale one was surprisingly slow and I don't seem to be able to eliminate it, but I moved it out of the hot loop.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12803 7d344553-34f0-0310-a9b1-970ce8f1c3a2