rasterizer: optimise Tex2DSampler/FPInt, which was killing GCC's optimiser
For me, this speeds up 8bit -> 32bit non-blended draws (of a backdrop) by ~10x
on x86_64 (GCC 10) and over 4x on x86, and gives a small 15% boost to clang too.

Speedup for blended blits about half that.

Surprisingly, converting to FPInts just for a couple lines inside the sampler is
much faster (up to 2x) than using floats there instead. Getting the floor of a
float is slower than I though.
I also tried converting TexCoord in the IncTypes to use FPInt to eliminate the
float->FPInt conversions, but bizarrely that was 2-3x slower.

Also, some commented-out failed optimisations to Color scaling

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12814 7d344553-34f0-0310-a9b1-970ce8f1c3a2