crashrpt now uses HTTPS, and then falls back to SMTP (which works!)
Also, immediately show the Name/Email and Description fields to encourage
users to fill them. And get rid of the possibly annoying "Send Later" popup
when you try to dismiss the crash report.

HTTPS fails on older versions of Windows such as XP.

Working SMTP requires either the new build of CrashSender1403.exe (previous
commit) which uses the To address also as the From field, or for the user to
input a valid email address.

If SMTP fails, SMAPI (opening the user's mail client) is used as the third option
(it used to be second).

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12879 7d344553-34f0-0310-a9b1-970ce8f1c3a2