Variable length slice extra data arrays
Slice.Extra is replaced with Slice.ExtraVec an integer vector which defaults to
NULL for length 3 arrays containing [0, 0, 0]. So the size of Slice is actually
reduced by this on average.

Added "resize slice extra", "slice extra length", "append slice extra"

Yes, HS has (primitive) arrays now!

I set the maximum length to 1000,000 because more than that is painfully slow
when loading or saving a slice collection, and I didn't want to have separate
limits in the editor and elsewhere

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12883 7d344553-34f0-0310-a9b1-970ce8f1c3a2