A new scheme for script handles: slice handles are no longer (quickly) reused
Script handles are int32s tagged with their type in the top 5 bits
so that different types of handles have distinct types.

Slice handles no longer count up from 1. They contain an 8-bit check ('ctr')
value so that even when a plotslices() slot is used a different slice handle
will be generated.

The number of slice handles (so by extension the number of slices) is now limited
to 2^21 = 2,097,152. The engine lags fairly badly if you have that many because
there's no way yet to disable AdvanceSlice. Now that we have slice extra data
arrays there's no need to create that many slices.

This means "slice is valid" no longer returns false positives if you pass it
an invalid handle or a previously freed handle, until at least 256*4000=1,024,000
slices are deleted.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12903 7d344553-34f0-0310-a9b1-970ce8f1c3a2