Fix enemy cursor positioning in battle (was broken after the switch to slice-based battle ui)
Also fixed cursor offset to actually match the position set in the enemy editor,
rather than being offset 3 pixels to the left for no good reason. This also means
that the default cursor position is the exact top center of the enemy sprite
instead of 3 pixels to the left of the exact top center.

This is technically a backcompat break, but since it is only 3 pixels, I don't want
to bother with a backcompat bit and I hope nobodywill notice except for feenick,
and I hope feenick will forgive me because it is more correct now!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12980 7d344553-34f0-0310-a9b1-970ce8f1c3a2