Require SSE2 on x86 (Pentium 4+) except Win95 nightlies. "scons sse2=0" to disable
This only affects Windows and Linux, not Mac or Android. Realistically, it only
matters for Win95 (music_sdl) builds, since gfx_sdl2 requires Windows XP anyway.

On x86, added a check at startup for whether SSE2 is present, and show an error
if not. FB actually does this automatically if you compile with -e, which we
don't.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13009 7d344553-34f0-0310-a9b1-970ce8f1c3a2