Work around GCC bug on Windows when using SSE2 in gengcc=0 (debug) builds
This resulted in a segmentation fault in C/C++ code (I saw it in
indexed_tree_from_palette) when a SSE2 instruction requiring alignment (movaps)
is called.

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=56597
GCC devs don't seem inclined to fix this bug rather than just requiring that if
you mix code generated by multiple Windows compilers you need to pass an extra
flag to GCC to tell it the stack isn't aligned.

I fear this is a problem in more cases than caught by this contrived 'if' condition.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13010 7d344553-34f0-0310-a9b1-970ce8f1c3a2