Add scons libdir=... arg & lookup SDL/SDL_mixer .so names at build time to use for dylibload
Also, use dylib_noload instead of dylibload to load SDL and SDL_mixer (and libmodplug)
on Unix.

verprint puts the exact filename of the libraries (e.g. libSDL2-2.0.so.0) in
globals.bas so they can be used to runtime link to the SDL/SDL_mixer .so file
we're actually linked to rather than one provided by the system (which provides
a libSDL2.so symlink to its copy).

scons libdir=... also is passed to the linker as a -L libpath

Also, the weird naming of libSDL/SDL_mixer in the commandgenius Android port is
now handled at build time, not with a special case in music_sdl.bas.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13059 7d344553-34f0-0310-a9b1-970ce8f1c3a2