Linux packages now include precompiled libraries and game/custom.sh wrappers
ohrrpgce-game/custom are now placed in linux/$arch/, and should be invoked
with the wrappers.
The ohrrpgce-player-linux package includes game.sh as-is, while the full package
renames game.sh to ohrrpgce-game and custom.sh to ohrrpgce-custom.

Added 'unix' target to ohrpackage.py which is same as 'linux' but doesn't
include the libraries and wrapper scripts. It's used for .deb packages and
"scons install".

distribmenu needs updating.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13072 7d344553-34f0-0310-a9b1-970ce8f1c3a2