SDL_mixer.dll recompiled for Win95 using gcc win32 threading model, fixes #1241
Using my mxe fork, commit cbafd6198036.

Also a couple fixes to build_sdl_mixer.sh to work in a clean mxe checkout.

Additionally I reverted the MIDI loop fix I backported for the previous build,
as it caused a freeze when stopping a MIDI on Win 95, but worked fine on WinXP.

    Versions used:
     GCC 5.5.0
     mingw-w64 7.0.0
     SDL_mixer git 3a4cdce6dca7 ("1.2.13")
     SDL 1.2.15
     libmodplug 0.8.9.0
     libogg 1.3.4
     libvorbis 1.3.7
     libmad 0.15.1b

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13108 7d344553-34f0-0310-a9b1-970ce8f1c3a2