distribmenu: rewrite/cleanup a lot of Windows packaging to future-proof it
On Windows when using an existing game.exe, extract buildinfo.ini from it to check it's an official sdl2 build.
If not, the user has the option to use it but is recommended to use a downloaded player instead.
The backends that Custom is built with and running are no long checked (yuck!), nor do
we check whether the game resolution is not 320x200 (yuck!), because all official builds have a
backend that can handle them.

When using a windows ohrrpgce-player zip, future-proof by including all files in it (except
buildinfo.ini and README-player-only.txt) instead of assuming it's a gfx_sdl2/music_sdl2
build. (buildinfo.ini isn't consulted.)  This is same way Linux packaging now works.
Innosetup also now includes all files in the temp directory.
And don't extract files into the _gameplayer dir, which was unlike all other ohrrpgce-player archives.

Replaced get_windows_gameplayer with add_windows_gameplayer, which directly copies all files into
the temp directory and renames game.exe.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13201 7d344553-34f0-0310-a9b1-970ce8f1c3a2