Work around FB bug sf#826 with separate SavePropBool[Always] functions
Without this workaround I doubt that the code is compiled correctly, since the
bug is that FB prefers function overloads for boolean rather than integer. But
that's not all, it seems to allow defining functions for boolean overloads that
haven't been declared, and which overload it picks seem to change after the
point in the file that it's defined.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13231 7d344553-34f0-0310-a9b1-970ce8f1c3a2