vec2Distance, vec3Distance, and some float2 (FB) & float3 (C++) overloaded operators
In particular, add XYPair.let(XYPair) so that you can write
  dim pos as XYPair
  pos = some_float2_expression
unfortunately there is no way in FB to make it possible to write
  dim pos as XYPair = some_float2_expression
unless we define XYPair before Float2 in util.bi which then means some of the
other operators can't be declared. This is because Operator Cast must be
declared inside the Type but can't use a forward-declared return type, and
we can't add constructors to XYPair because it's used in unions.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13250 7d344553-34f0-0310-a9b1-970ce8f1c3a2