Tile/spriteedit: when picking a color by right-click/G/Enter, change to Draw tool if using a non-drawing tool (mainly Scroll)
This sounds like an unimportant tweak but I probably accidentally scrolled 40
times today because it didn't do this!

(I should make it do the same when you change the selected color in other ways too)

You can no longer use Scroll in the tile editor by right-dragging. You couldn't
in the sprite editor. Also, G and ENTER to pick a color now work in the sprite
editor when you Mark or Clone selected

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13277 7d344553-34f0-0310-a9b1-970ce8f1c3a2