scons: fix a transpiledir= race condition
It seems scons runs Actions in parallel so sometimes a file Copy was run
before the destination directory's Mkdir.

Half-rewrote ohrbuild.copy_source_actions yet again to use a custom Action, but
in future it can be deleted if we make SCons 4.5.0 a requirement.

This also reduces the file copying spam.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13446 7d344553-34f0-0310-a9b1-970ce8f1c3a2