Small graphical fix for multichoice wiping contents of vpage
This was also caused by r13411. If you call multichoice twice in a row (notable
"Really quit?" prompt in the quit menu) the first call showed the old screen
contents (holdpage) underneath, but didn't restore vpage after due to the
8->32->8 bit switch, so the 2nd menu was black underneath.

There may be other menus (show_help?) with the same glitch but it was mainly
just the quit menu that bothered me.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13495 7d344553-34f0-0310-a9b1-970ce8f1c3a2