hspeak: Add debug info to compiled scripts
Adds a srcfiles.txt lump which is used to decode srcpos (to go from source offset to file/line/column).

Each .hsz gains an extended header, a table of local variable names, and each
tyscript, tymath, tyfunct and tyflow node is appended by it srcpos. This is
backward compatible with old script interpreters.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13501 7d344553-34f0-0310-a9b1-970ce8f1c3a2