Replace anim_flinchdone with a timer triggered by anim_flinchstart
The BattleSprite.flinch_anim timer causes flinch_anim_eachtick to be called
to handle the slide back to initial position animation after 3 ticks.

The reason for this change is to allow changing the lengths of waits in the
attack animations without upsetting the flinch animation. For example
Sequential Projectile pauses on each target for 3 ticks because it's waiting
for the flinch.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13551 7d344553-34f0-0310-a9b1-970ce8f1c3a2